backend/pyproject.toml
[project]
name = "rag-knowledge-ingestion"
version = "0.1.0"
description = "RAG Knowledge Ingestion Pipeline for Docusaurus Book"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "cohere>=4.4.3",
    "qdrant-client>=1.9.1",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "fastapi[standard]>=0.124.4",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "openai>=2.12.0",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
